:art: Applied PR suggestions
After submitting the PR, I've received some feedbacks regarding the code.

On this commit I'm applying those feedbacks.

Now, our code should be working as expected.
