+++ cmd/kubeless/function/deploy.go
@@ -26,7 +26,7 @@ import (
 	cronjobUtils "github.com/kubeless/cronjob-trigger/pkg/utils"
 	kubelessApi "github.com/kubeless/kubeless/pkg/apis/kubeless/v1beta1"
 	"github.com/kubeless/kubeless/pkg/langruntime"
-	kubelessUtils "github.com/kubeless/kubeless/pkg/utils"
+	kubelessutil "github.com/kubeless/kubeless/pkg/utils"
 	"github.com/robfig/cron"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
@@ -38,8 +38,8 @@ var deployCmd = &cobra.Command{
 	Short: "deploy a function to Kubeless",
 	Long:  `deploy a function to Kubeless`,
 	Run: func(cmd *cobra.Command, args []string) {
-		cli := kubelessUtils.GetClientOutOfCluster()
-		apiExtensionsClientset := kubelessUtils.GetAPIExtensionsClientOutOfCluster()
+		cli := kubelessutil.GetClientOutOfCluster()
+		apiExtensionsClientset := kubelessutil.GetAPIExtensionsClientOutOfCluster()
 
 		if len(args) != 1 {
 			logrus.Fatal("Need exactly one argument - function name")
@@ -52,7 +52,7 @@ var deployCmd = &cobra.Command{
 		}
 
 		// Checking runtime parameter if allowed by RBAC, otherwide skip the check
-		config, err := kubelessUtils.GetKubelessConfig(cli, apiExtensionsClientset)
+		config, err := kubelessutil.GetKubelessConfig(cli, apiExtensionsClientset)
 		if config == nil || err != nil {
 			logrus.Warnf("%v. Runtime check is disabled.", err)
 		} else {
@@ -101,7 +101,7 @@ var deployCmd = &cobra.Command{
 			logrus.Fatal(err)
 		}
 		if ns == "" {
-			ns = kubelessUtils.GetDefaultNamespace()
+			ns = kubelessutil.GetDefaultNamespace()
 		}
 
 		deps, err := cmd.Flags().GetString("dependencies")
@@ -177,11 +177,11 @@ var deployCmd = &cobra.Command{
 
 		funcDeps := ""
 		if deps != "" {
-			contentType, err := getContentType(deps)
+			contentType, err := kubelessutil.GetContentType(deps)
 			if err != nil {
 				logrus.Fatal(err)
 			}
-			funcDeps, _, err = parseContent(deps, contentType)
+			funcDeps, _, err = kubelessutil.ParseContent(deps, contentType)
 			if err != nil {
 				logrus.Fatal(err)
 			}
@@ -227,13 +227,13 @@ var deployCmd = &cobra.Command{
 			}
 		}
 
-		kubelessClient, err := kubelessUtils.GetKubelessClientOutCluster()
+		kubelessClient, err := kubelessutil.GetKubelessClientOutCluster()
 		if err != nil {
 			logrus.Fatal(err)
 		}
 
 		logrus.Infof("Deploying function...")
-		err = kubelessUtils.CreateFunctionCustomResource(kubelessClient, f)
+		err = kubelessutil.CreateFunctionCustomResource(kubelessClient, f)
 		if err != nil {
 			logrus.Fatalf("Failed to deploy %s. Received:\n%s", funcName, err)
 		}
