+++ cmd/kubeless/trigger/cronjob/update.go
@@ -72,6 +72,21 @@ var updateCmd = &cobra.Command{
 			logrus.Fatal(err)
 		}
 
+		payload, err := cmd.Flags().GetString("payload")
+		if err != nil {
+			logrus.Fatal(err)
+		}
+
+		payloadFromFile, err := cmd.Flags().GetString("payload-from-file")
+		if err != nil {
+			logrus.Fatal(err)
+		}
+
+		if len(payload) > 0 && len(payloadFromFile) > 0 {
+			err := "You can't provide both raw payload and a payload file"
+			logrus.Fatal(err)
+		}
+
 		kubelessClient, err := kubelessUtils.GetKubelessClientOutCluster()
 		if err != nil {
 			logrus.Fatalf("Can not create out-of-cluster client: %v", err)
@@ -87,12 +102,18 @@ var updateCmd = &cobra.Command{
 			logrus.Fatalf("Unable to find Function %s in namespace %s. Error %s", triggerName, ns, err)
 		}
 
+		parsedPayload, err := parsePayload(payload, payloadFromFile)
+		if err != nil {
+			logrus.Fatal("Unable to parse the payload of Function %s in namespace %s. Error %s", functionName, ns, err)
+		}
+
 		cronJobTrigger, err := cronjobUtils.GetCronJobCustomResource(cronJobClient, triggerName, ns)
 		if err != nil {
 			logrus.Fatalf("Unable to find Cronjob trigger %s in namespace %s. Error %s", triggerName, ns, err)
 		}
 		cronJobTrigger.Spec.FunctionName = functionName
 		cronJobTrigger.Spec.Schedule = schedule
+		cronJobTrigger.Spec.Payload = parsedPayload
 
 		if dryrun == true {
 			res, err := kubelessUtils.DryRunFmt(output, cronJobTrigger)
@@ -117,4 +138,6 @@ func init() {
 	updateCmd.Flags().StringP("function", "", "", "Name of the function to be associated with trigger")
 	updateCmd.Flags().Bool("dryrun", false, "Output JSON manifest of the function without creating it")
 	updateCmd.Flags().StringP("output", "o", "yaml", "Output format")
+	updateCmd.Flags().StringP("payload", "p", "", "Specify a stringified JSON data to pass to function upon execution")
+	updateCmd.Flags().StringP("payload-from-file", "f", "", "Specify a payload file to use. It must be a JSON file")
 }
