+++ pkg/utils/kubelessutil.go
@@ -22,11 +22,15 @@ import (
 	"encoding/json"
 	"fmt"
 	"os"
+	"io"
+	"io/ioutil"
 	"path"
 	"strconv"
 	"strings"
-
-	"io/ioutil"
+	"net/url"
+	"net/http"
+	"encoding/base64"
+	"unicode/utf8"
 
 	monitoringv1alpha1 "github.com/coreos/prometheus-operator/pkg/client/monitoring/v1alpha1"
 	"github.com/ghodss/yaml"
@@ -873,4 +877,104 @@ func DryRunFmt(format string, trigger interface{}) (string, error) {
 	default:
 		return "", fmt.Errorf("Output format needs to be yaml or json")
 	}
+}
+
+// Gets the content type of a given filename
+func GetContentType(filename string) (string, error) {
+	var contentType string
+
+	if strings.Index(filename, "http://") == 0 || strings.Index(filename, "https://") == 0 {
+		contentType = "url"
+		if path.Ext(strings.Split(filename, "?")[0]) == ".zip" {
+			contentType += "+zip"
+		}
+	} else {
+		fbytes, err := ioutil.ReadFile(filename)
+		if err != nil {
+			return "", err
+		}
+		isText := utf8.ValidString(string(fbytes))
+		if isText {
+			contentType = "text"
+		} else {
+			contentType = "base64"
+			if path.Ext(filename) == ".zip" {
+				contentType += "+zip"
+			}
+		}
+	}
+	return contentType, nil
+}
+
+// Parse the content of a file as string
+func ParseContent(file, contentType string) (string, string, error) {
+	var checksum, content string
+
+	if strings.Contains(contentType, "url") {
+
+		functionURL, err := url.Parse(file)
+		if err != nil {
+			return "", "", err
+		}
+		resp, err := http.Get(functionURL.String())
+		if err != nil {
+			return "", "", err
+		}
+		defer resp.Body.Close()
+
+		functionBytes, err := ioutil.ReadAll(resp.Body)
+		if err != nil {
+			return "", "", err
+		}
+		content = string(functionBytes)
+		checksum, err = getSha256(functionBytes)
+		if err != nil {
+			return "", "", err
+		}
+
+	} else {
+
+		functionBytes, err := ioutil.ReadFile(file)
+		if err != nil {
+			return "", "", err
+		}
+		if contentType == "text" {
+			content = string(functionBytes)
+		} else {
+			content = base64.StdEncoding.EncodeToString(functionBytes)
+		}
+		checksum, err = getFileSha256(file)
+		if err != nil {
+			return "", "", err
+		}
+	}
+
+	return content, checksum, nil
+}
+
+// Get the checksum of a file using sha256
+func getFileSha256(file string) (string, error) {
+	h := sha256.New()
+	ff, err := os.Open(file)
+	if err != nil {
+		return "", err
+	}
+	defer ff.Close()
+	_, err = io.Copy(h, ff)
+	if err != nil {
+		return "", err
+	}
+	checksum := hex.EncodeToString(h.Sum(nil))
+	return "sha256:" + checksum, err
+}
+
+// Get the checksum using sha256
+func getSha256(bytes []byte) (string, error) {
+	h := sha256.New()
+	_, err := h.Write(bytes)
+	if err != nil {
+		return "", err
+	}
+	checksum := hex.EncodeToString(h.Sum(nil))
+	return "sha256:" + checksum, nil
 }
